refactor: move prepared dml arena below session
